Removed Unicode Artifacts from Ready.txt
When I wrote this story in Google Docs, some of the characters were
automatically replaced with Unicode for better formatting. When I
copy/pasted to a .txt file it kept some of those. Unfortunately, .txt
files don't display anything but ASCII correctly in a web browser.

I should keep an eye out for other Unicode artifacts.