Moved "A Message of Hope" to the serials folder
Decided "A Message of Hope will eventually be a series, moved it to the
serials folder, and silently updated the entry in the index.

When I post the second story in the series, I'll have to update the
entry in the index to point to a folder rather than a text file.