Reframe Logan stories + add to GamePermissions.html
I decided to change the framing device for the Logan story and put it in
a personnel file instead. This allows a lot more freedom, makes it so I
don't have to have KAI analyze everything (but he still could), and even
makes it so that if LRL ever gets more than one contributor on this
branch, I can add personnel files for them as well.

Also added Fallen London and Sunless Sea to GamePermissions.html and
removed the commented out link from the landing page.