Rewrite to use an implied index.html where possible
This way the user doesn't have to type the name of the webpage or even
see it in the URL bar. Curiously enough, about half of the pages that
would be an index.html file were already.

This may break some existing links from other websites, but before I
push this to the live server I'll be writing a custom 404 page.

Also, this makes testing using file:// difficult when navigating between
folders. Using a localhost server fixes that, but be prepared, and don't
forget to test the live site when these changes are merged.