Mouse-over effects, plus clean-up
Made it so that links change color when you hover over them, and set the
site to do that effect briefly upon loading the page.

Also cleaned up a little code, like using a separate .js file for Google
Analytics, rewriting navigation links to use absolute paths, and
rebounding links in the header.